+++ packages/peekaboo/src/react/usePeekaboo.ts
@@ -1,15 +1,30 @@
 'use client';
 
-import { useEffect, useReducer, useRef } from 'react';
+import { useEffect, useReducer } from 'react';
 import { BooType, UpdateDetail } from '../types';
 import { UPDATE_VALUE, INIT_VALUE } from '../consts';
 import { validateBoo } from '../utils';
 
+// const deepEquial = <T>(a: T, b: T) => {
+
+// 	if (a === b) return true;
+// 	if (typeof a !== 'object' || typeof b !== 'object') return false;
+// 	const keysA = Object.keys(a);
+// 	const keysB = Object.keys(b);
+// 	if (keysA.length !== keysB.length) return false;
+// 	for (let key of keysA) {
+// 		if (!keysB.includes(key) || !deepEquial(a[key], b[key])) return false;
+// 	}
+// 	return true;
+// }
+
 export const usePeekaboo = <T>(boo: BooType<T>) => {
 	validateBoo(boo);
-	const [state, setState] = useReducer((s: T, a: (_v: T) => T) => a(s), boo.get() ?? (boo.init() as T));
-	const valRef = useRef<T>(state);
-	valRef.current = state;
+	const [state, setState] = useReducer((s: { value: T }, a: (_v: { value: T }) => { value: T }) => a(s), {
+		value: boo.get() ?? (boo.init() as T),
+	});
+	// const valRef = useRef<T>(state);
+	// valRef.current = state;
 
 	useEffect(() => {
 		const listener: EventListenerOrEventListenerObject = e => {
@@ -21,13 +36,18 @@ export const usePeekaboo = <T>(boo: BooType<T>) => {
 			const shouldInit = ev.type === INIT_VALUE && ev.detail.storeId === boo.__store.storeId;
 
 			if (shouldUpdate || shouldInit) {
-				if (valRef.current === boo.get() && !ev.detail.forceRender) return;
+				const updated = boo.get() as T;
+
+				// if (valRef.current === updated && !ev.detail.forceRender) return;
 				setState(state => {
-					if (ev.detail.forceRender && typeof boo.get() === 'object') {
-						return { ...boo.get() } as T;
+					if (ev.detail.forceRender && ev.detail.forceRender && typeof updated === 'object') {
+						return { value: { ...updated } };
 					}
-					if (boo.get() !== state || ev.detail.forceRender) return boo.get() as T;
+					if (state.value !== updated) {
+						return { value: updated };
+					}
 					return state;
+					// return state;
 				});
 			}
 		};
@@ -39,5 +59,5 @@ export const usePeekaboo = <T>(boo: BooType<T>) => {
 		};
 	}, [boo]);
 
-	return state;
+	return state.value;
 };