+++ apps/demo/src/app/page.tsx
@@ -1,13 +1,5 @@
-'use client';
+import React from 'react';
 
-import React, { PropsWithChildren } from 'react';
-import Updated from '../components/Updated';
-import { peekaboo } from '../const';
-import Trigger from '../components/Trigger';
-import Link from 'next/link';
-
-export default function Page({ children }: PropsWithChildren) {
-	// const dropDownIndex = useAtomValue(peeks.atoms.page1.dropDownIndex.init());
-
-	return <div>{children}</div>;
+export default function Page() {
+	return <div>Main</div>;
 }