+++ packages/peekaboo/CHANGELOG.md
@@ -15,3 +15,4 @@ add size-limit
 
 ## 1.3.0
 add everUsed function
+make all mid level structure Boo