+++ packages/peekaboo/CHANGELOG.md
@@ -1,19 +1,33 @@
 # Change Logs
 
+## 1.4.0
+
+### Minor Changes
+
+- 998662b: peekaboo-store works with react and vanilla-js
+
+### Patch Changes
+
+- e34c0a9: core is working
+
 ## 1.1.0
+
 first version (prev versions are removed)
 support better slice type,
 restrict initial data for the easier usage
 
 ## 1.1.1
+
 type mismatch support
 add warning messages for type mismatch, does not exist update
 
 ## 1.2.0 (semi major update)
+
 split utils
 add size-limit
 
 ## 1.3.0
+
 add everUsed function
 make all mid level structure Boo
 adjust object for better typescript usage.